(y60) -fixed destination pixel format - reinitialize lastVideoFrame
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@24590 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ y60/video/Movie.cpp
@@ -181,7 +181,7 @@ namespace y60 {
             break;
         case PLAY_MODE_PLAY:
             if (_myPlayMode == PLAY_MODE_PAUSE) {
-                if (get<CurrentFrameTag>() == _myLastDecodedFrame) {
+                if (get<CurrentFrameTag>() == static_cast<int>(_myLastDecodedFrame)) {
                     _myDecoder->resumeMovie(getTimeFromFrame(get<CurrentFrameTag>()));
                 } else {
                     _myDecoder->startMovie(getTimeFromFrame(get<CurrentFrameTag>()));
